+++ SisInfInt-Grupo-C-ejb/src/test/java/es/uma/informatica/jpa/demo/ejb/practica/BaseDatos.java
@@ -54,7 +54,7 @@ public class BaseDatos {
 
 		final Integer numExpediente = 11;
 		final Boolean activo = true;
-		final Integer notaMediaProvisional = 9;
+		final Double notaMediaProvisional = 9.0;
 		
 		Expediente exp1 = new Expediente(numExpediente, activo, notaMediaProvisional);
 		em.persist(exp1);
